Ignore mousewheel events for motif sliders CATER 118737 - SLAC Local patch
Some of the magnet control screens are large and have a scrollbar
When using the mouse wheel to scroll down the page, we sometimes accidentally scroll over motifSliders
In this case, the motif slider captures the mouse wheel events and changes random magnet controls
Request from ops was to ignore mouse wheel events for MOTIF sliders.
This local patch request was requested by Mike Stanek, Matt Gibbs and Mike Zelazny
Ernest Williams and Mike Zelazny approved the local patch
Murali Shankar made this local patch on Aug/12/2014.
We are not sending this patch to the collaboration as EDM has another slider from Triumf that sorta addresses this.
However, it was determined that the TRIUMF slider does not fit our needs.